requirements.txt
streamlit
openai
google.oauth2
googleapiclient.discovery
googleapiclient.http